pyproject.toml
[project]
name = "sprint-03-menu-generator"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.11.0",
    "diffusers>=0.35.2",
    "googletrans==4.0.0rc1",
    "pillow>=12.0.0",
    "safetensors>=0.6.2",
    "transformers>=4.57.1",
]
